// Quellcode/src/app/Model/sterbefaelle.ts
export class Sterbefaelle {
  id: number | null = null;
  jahr: number | null = null;
  gesamt: number | null = null;
  januar: number | null = null;
  februar: number | null = null;
  maerz: number | null = null;
  april: number | null = null;
  mai: number | null = null;
  juni: number | null = null;
  juli: number | null = null;
  august: number | null = null;
  september: number | null = null;
  oktober: number | null = null;
  november: number | null = null;
  dezember: number | null = null;
}